requirements.txt
ansible-vault~=1.2.0
Click~=7.0
PyYAML~=5.3
python-editor~=1.0.4
